app/f_test/requirements.txt
Django>=1.8,<2.0
gunicorn==19.7.1
django-ipware==1.1.6
